backend/requirements.txt
Django==4.1.1
djangorestframework==3.13.1
django-filter==22.1
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
djoser==2.1.0